fix(project_and_component): Removed code duplication of ShowInfoOnHover and made it global component
