+++ src/components/ProjectAddSummary/component/Administration/Clearing.tsx
@@ -9,8 +9,7 @@
 
 'use client'
 
-import { OverlayTrigger, Tooltip } from 'react-bootstrap'
-import { BiInfoCircle } from 'react-icons/bi'
+import ShowInfoOnHover from '@/components/ShowInfoOnHover/ShowInfoOnHover'
 import { useTranslations } from 'next-intl'
 import { COMMON_NAMESPACE } from '@/object-types/Constants'
 import ProjectPayload from '@/object-types/CreateProjectPayload'
@@ -22,18 +21,6 @@ interface Props{
     setProjectPayload: React.Dispatch<React.SetStateAction<ProjectPayload>>
 }
 
-
-const ShowInfoOnHover = ({ text }: { text: string }) => {
-    return (
-        <>
-            <OverlayTrigger overlay={<Tooltip>{text}</Tooltip>}>
-                <span className='d-inline-block'>
-                    <BiInfoCircle />
-                </span>
-            </OverlayTrigger>
-        </>
-    )
-}
 
 export default function Clearing({token, projectPayload, setProjectPayload}: Props) {
